==> 0_get_data.ipynb <==
import pandas as pd
import warnings
from pprint import pprint

from src import wiki, news, twitter
warnings.filterwarnings("ignore")

# Wikipedia

Retrieve Wikipedia pages that have titles matching a particular topic of interest, and then extract the relevant sections from those pages.

Next, examine all the articles referenced within these pages. If a referenced article is relevant to the topic of interest and contains specific keywords, retrieve the pages of those articles as well.


# Extract sections from Wikipedia articles for the listed titles.
# And recursiverly extract all linked articles, filter given the keywords.

df_wiki = wiki.get_wiki_pages(titles = ["2022 FIFA World Cup"], 
                         keywords = ["2022", "fifa", "world", "cup"])

Examples of titles/sections extracted:

df_wiki.head(10)

Example of one section:

idx = 3
print("title: ", df_wiki.title.iloc[idx])
print("heading: ", df_wiki.heading.iloc[idx])
print("text: ", df_wiki.text.iloc[idx])

print("Examples of extracted wiki pages titles:")
print(*("- " + df_wiki.title.unique()[:10]), sep='\n')


# News API

Use News REST API to get the list of news articles published about a certain topic.

The free subscription allows searching articles up to a month old only.

The News API gets the unformatted content of the article, where available. This is truncated to 200 chars. We can then use `newspaper` python library to extract all content give article's url.

# News API: Extract News article regarding a certain topic.
# Free API is limited to last 30 days.

NEWS_API_KEY = ""

query = "World Cup 2022"
from_date = ""
to_date = ""

sort_by = "popularity" # popularity, relevancy, publishedAt
searchIn = "title" # match query in title, description, content (can be multiple)
page_size = 20

df_news = news.get_news_pages(NEWS_API_KEY, query, from_date, to_date, 
                            sort_by, searchIn, page_size)


df_news.head()

print("Examples of extracted wiki pages titles:")
print(*("- " + df_news.title.unique()[:10]), sep='\n')

Not all titles are relevant to our topic. Event happened 3 month, free api limited to last month.

idx = 2
print("title: ", df_news.title.iloc[idx])
print("author: ", df_news.author.iloc[idx])
print("description: ", df_news.description.iloc[idx])
print("url: ", df_news.url.iloc[idx])
print("text: ", df_news.text.iloc[idx])

# Twitter

# twitter api is limited to last 7 days

TWITTER_API_KEY = ""
TWITTER_API_SECRET_KEY = ""
TWITTER_ACCESS_TOKEN = ""
TWITTER_ACCESS_TOKEN_SECRET = ""

query = "Fifa World Cup"
language = "en"
num_tweets = 1000
include_RTs = False

auth_api = twitter.twitter_api_auth(TWITTER_API_KEY, TWITTER_API_SECRET_KEY, TWITTER_ACCESS_TOKEN, TWITTER_ACCESS_TOKEN_SECRET)
df_twitter = twitter.get_tweets(query, language, num_tweets, auth_api, include_RTs)


# ALL DATA

Concatenate all data content together in one dataframe

df = pd.concat([df_wiki, df_news, df_twitter],join='inner', ignore_index=True)
df.head()


df['source'].value_counts()

sources = [
    "wiki",
    #"news", 
    #"twitter"
]

df = df.loc[df["source"].isin(sources)]


Number of Sections in data:

len(df)

df.to_csv("data/fifa_data.csv")
