==> 1_prepare_qa_finetuning.ipynb <==
import pandas as pd
import openai
from src import wiki, news, twitter
import os

import warnings
from pprint import pprint

warnings.filterwarnings("ignore")

# Azure OpenAI 
# Insert your API endpoint URL & key
openai.api_type = "azure"
openai.api_version = "2022-12-01"
openai.api_base = ""
openai.api_key = ""

# Read Data

df = pd.read_csv("data/fifa_data.csv", index_col=0)
df = df.iloc[:10]
df.head()

# Generate Questions 

def get_questions(context):
    try:
        response = openai.Completion.create(
            engine="text-davinci-003-base",
            prompt=f"Write all possible questions that can be answered from the text below. \n\nText: {context}\n\nQuestions:\n1.",
            temperature=0,
            max_tokens=400,
            top_p=1,
            frequency_penalty=0,
            presence_penalty=0,
            stop=["\n\n"]
        )
        return response['choices'][0]['text']
    except:
        return ""


df['questions']= df.text.apply(get_questions)
df['questions'] = "1." + df.questions

print(df["text"].iloc[4])

# Generated Questions Example:

print(df["questions"].iloc[4])

# Get Answers

Get answers based on context.

def get_answers(row):
    try:
        response = openai.Completion.create(
            engine="text-davinci-003-base",
            prompt=f"Write answers to the questions below based on the context given in the text below. \n\nText: {row.text}\n\nQuestions:\n{row.questions}\n\nAnswers:\n1.",
            temperature=0,
            max_tokens=500,
            top_p=1,
            frequency_penalty=0,
            presence_penalty=0
        )
        return response['choices'][0]['text']
    except Exception as e:
        print (e)
        return ""


df['answers'] = df.apply(get_answers, axis=1)
df['answers'] = "1." + df.answers
df = df.dropna().reset_index().drop('index',axis=1)

# Questions:

print(df["questions"].iloc[4])

# Answers:

print(df["answers"].iloc[4])

# Prepare Q&A

Transform questions & answers from numbered bullets into python lists.

import re

def split_text(text):
    return([x for x in re.split(r'((?:^\s*\d+\. )|(?:\s+\d+\. ))\s*', 
                            text) if x and not re.match(r"\s*\d+\. ",x)])

df["questions"] = df.questions.apply(split_text)
df["answers"] = df.answers.apply(split_text)

Transform dataframe into long format where each row is one question & answer.

df.iloc[4].questions

df["qa_match"] = df.apply(lambda x: len(x["questions"])==len(x["answers"]), axis=1)
df = df.loc[df['qa_match'] == True]
df = df.drop(['qa_match'], axis=1)
df.head()

df.head(11)

df_long = pd.DataFrame()
for index, cols in df.iterrows():
    df_extract = pd.DataFrame.from_dict({'questions': cols['questions'], 'answers': cols['answers']})
    df_long = pd.concat([df_long, df_extract], ignore_index=True)

df_long.rename(columns={"questions": "prompt", "answers": "completion"}, inplace=True)

df_long.tail()

Number of questions & answers in the document

len(df_long)

# Save generated Q&A

df_long.to_json('fifa-data/fifa_qa_gen.jsonl', orient="records", lines=True)
