==> 2_finetuning_q&a.ipynb <==
# Connecting to Azure Open AI

import os
import openai

# Azure OpenAI 
# Insert your API endpoint URL & key
openai.api_type = "azure"
openai.api_version = "2022-12-01"
openai.api_base = ""
openai.api_key = ""

!openai tools fine_tunes.prepare_data -f fifa-data/fifa_qa.jsonl -q

!openai api fine_tunes.create -t "fifa-data/fifa_qa_prepared.jsonl" -m davinci --suffix "world cup 2022"

!openai api fine_tunes.follow -i ft-U98OejdQvjYzc6x5wjKAijts

# Use Base Model

model = 'text-davinci-003'
res = openai.Completion.create(model=model, prompt="Who won Fifa World Cup 2022?", max_tokens=100, temperature=0)
res['choices'][0]['text']

# Use Finetuned Model

ft_model = 'davinci:ft-microsoft:world-cup-2022-2023-02-09-17-27-13'
res = openai.Completion.create(model=ft_model, prompt="Who won Fifa World Cup 2022?", max_tokens=1, temperature=0)
res['choices'][0]['text']
